requirements.txt
aiodns==3.2.0
aiosmtpd==1.4.5
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
atpublic==4.1.0
attrs==23.2.0
backports.zoneinfo==0.2.1
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
diff-match-patch==20230430
Django==4.2.13
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-import-export==4.0.3
django-redis==5.4.0
django-rq==2.10.2
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
gunicorn==22.0.0
numpy==1.24.4
packaging==24.1
paramiko==3.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycares==4.4.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
rq==1.16.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
times==0.7
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.2
tzdata==2024.1
verify-email==2.4.3
